fix: upgrade detect version
