+++ src/components/RotateScreen/RotateScreen.tsx
@@ -1,29 +1,29 @@
-import { memo } from 'react';
+import { memo, useEffect, useState } from 'react';
 import classnames from 'classnames';
-import { useOrientation } from 'react-use';
+import { device } from '@jam3/detect';
 
 import styles from './RotateScreen.module.scss';
-
-import detect from '../../utils/detect';
 
 export type Props = {
   className?: string;
 };
 
 function RotateScreen({ className }: Props) {
-  const orientation = useOrientation();
+  const [enable, setEnable] = useState(device.phone && device.landscape);
 
-  return (
-    <div
-      className={classnames(
-        styles.RotateScreen,
-        { [styles.enabled]: detect.device.isPhone && orientation.angle !== 0 },
-        className
-      )}
-    >
-      Rotate Screen
-    </div>
-  );
+  useEffect(() => {
+    const handleResize = () => {
+      setEnable(device.phone && device.landscape);
+    };
+
+    window.addEventListener('resize', handleResize);
+
+    return () => {
+      window.removeEventListener('resize', handleResize);
+    };
+  }, []);
+
+  return <div className={classnames(styles.RotateScreen, { [styles.enabled]: enable }, className)}>Rotate Screen</div>;
 }
 
 export default memo(RotateScreen);
