--- src/utils/detect.ts
@@ -1,42 +1,0 @@
-import noop from 'no-op';
-
-const isBrowser = typeof window !== 'undefined';
-
-const detect = {
-  device: { isDesktop: true, isMobile: false, isPhone: false, getType: noop },
-  browser: { getName: noop },
-  os: {},
-  bots: {}
-};
-
-if (isBrowser) {
-  detect.device = require('@jam3/detect').device;
-  detect.browser = require('@jam3/detect').browser;
-  detect.os = require('@jam3/detect').os;
-  detect.bots = require('@jam3/detect').bots;
-}
-
-export const isTouchDevice = isBrowser && ('ontouchstart' in window || navigator.maxTouchPoints > 0);
-
-const testImages = {
-  lossy: 'UklGRiIAAABXRUJQVlA4IBYAAAAwAQCdASoBAAEADsD+JaQAA3AAAAAA',
-  lossless: 'UklGRhoAAABXRUJQVlA4TA0AAAAvAAAAEAcQERGIiP4HAA==',
-  alpha:
-    'UklGRkoAAABXRUJQVlA4WAoAAAAQAAAAAAAAAAAAQUxQSAwAAAARBxAR/Q9ERP8DAABWUDggGAAAABQBAJ0BKgEAAQAAAP4AAA3AAP7mtQAAAA==',
-  animation:
-    'UklGRlIAAABXRUJQVlA4WAoAAAASAAAAAAAAAAAAQU5JTQYAAAD/////AABBTk1GJgAAAAAAAAAAAAAAAAAAAGQAAABWUDhMDQAAAC8AAAAQBxAREYiI/gcA'
-};
-
-export function checkWebpSupport(feature: keyof typeof testImages, callback: (isSupported: boolean) => void): void {
-  var img = new Image();
-  img.onload = function () {
-    var result = img.width > 0 && img.height > 0;
-    callback(result);
-  };
-  img.onerror = function () {
-    callback(false);
-  };
-  img.src = 'data:image/webp;base64,' + testImages[feature];
-}
-
-export default detect;
